requirements.txt
aiohttp==3.11.4
aiosignal==1.3.2
attrs==25.3.0
contourpy==1.3.1
croniter==6.0.0
cycler==0.12.1
discord-py==2.5.2
discord-typings==0.9.0
emoji==2.11.1
fonttools==4.56.0
frozenlist==1.5.0
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.2.0
numpy==2.2.0
packaging==24.2
pillow==11.1.0
pilmoji==2.0.4
pre-commit
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tomli==2.2.1
typing_extensions==4.13.0
yarl==1.18.3
